requirements.txt
annotated-types==0.7.0
anyio==4.11.0
apache-beam==2.48.0
avro-python3==1.10.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.3.0
cloudpickle==2.2.1
confluent-kafka==2.12.1
crcmod==1.7
dataclasses==0.6
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
fastapi==0.119.0
fastavro==1.10.0
fasteners==0.19
find_libpython==0.5.0
grpcio==1.70.0
h11==0.16.0
hdfs==2.7.3
httplib2==0.22.0
ibm_db==3.2.7
idna==3.10
numpy==1.24.4
objsize==0.6.1
orjson==3.10.15
pandas==2.2.3
pemja==0.4.1
proto-plus==1.26.0
protobuf==4.23.4
py4j==0.10.9.7
pyaml==25.1.0
pyarrow==11.0.0
pydantic==2.11.10
pydantic_core==2.33.2
pydot==1.4.2
pyflink==1.0
pymongo==4.11.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.1
PyYAML==6.0.2
redis==7.0.0
regex==2024.11.6
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.37.0
zstandard==0.23.0
